feat: added ability to filter databases
