+++ src/data_hub/api.py
@@ -873,6 +873,7 @@ def get_dataframe(
     use_file_names: bool = True,
     reference_format: IDFormat = "human-id",
     return_type: DatabaseReturnType = "dataframe",
+    filter: Optional[dict] = None,
     client=None,
 ):
     """Generate a `pandas.DataFrame` or dictionary for a database.
@@ -890,10 +891,17 @@ def get_dataframe(
     # TODO: list_database_rows and describe_row can be called in parallel
 
     # figure out the rows
-    rows = _api.list_database_rows(
-        database_row_id=database_id,
-        client=client,
-    )
+    if filter is None:
+        rows = _api.list_database_rows(
+            database_row_id=database_id,
+            client=client,
+        )
+    else:
+        rows = _api.list_database_rows(
+            database_row_id=database_id,
+            client=client,
+            filter=filter,
+        )
 
     # figure out the column names and ID of the database
     db_row = _api.describe_row(
