+++ src/data_hub/dataframe.py
@@ -9,6 +9,7 @@ from datetime import datetime, timezone
 import humanize
 import pandas as pd
 from beartype import beartype
+from beartype.typing import Optional
 from dateutil.parser import parse
 from deeporigin.data_hub import api
 from deeporigin.platform.api import get_last_edited_user_name
@@ -263,6 +264,7 @@ class DataFrame(pd.DataFrame):
         use_file_names: bool = True,
         reference_format: IDFormat = "human-id",
         client=None,
+        filter: Optional[dict] = None,
     ):
         """Create a local Deep Origin DataFrame from a Deep Origin database.
