+++ src/utils/types.py
@@ -1,44 +1,7 @@
 """this module contains types used in the underlying SDK,
 that are then re-exposed for easier use."""
 
-from beartype import beartype
 from deeporigin_data import types
-from deeporigin_data.types.client_list_database_rows_params import FilterRowFilterNumber
-from pydantic import TypeAdapter
 
 ListFilesResponse = types.list_files_response.Data
 DataFile = types.list_files_response.DataFile
-
-
-@beartype
-def numeric_condition(
-    *,
-    column_id: str,
-    value: float | int,
-    operator: str,
-) -> dict:
-    """function to create a valid numeric filter
-
-
-    Args:
-        column_id (str): id of the column
-        value (float | int): value to filter
-        operator (str): operator to use.
-
-    Returns:
-        A dictionary containing a valid condition that can be used to filter database rows
-    """
-
-    data = dict(
-        column_id=column_id,
-        filter_type="number",
-        filter_value=value,
-        operator=operator,
-    )
-
-    # use type checks from SDK definition, which
-    # ultimately derives from the openAPI spec
-    adapter = TypeAdapter(FilterRowFilterNumber)
-    adapter.validate_python(data)
-
-    return data
